+++ Runtime/MRExtensionMethods_GameObjects.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6c71ffd37fe8b4a09858c7b4807d1329
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
