+++ Runtime/MRExtensionMethods_IEnumerators.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7725e3105e71c4f479f976e3f32d956f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
