+++ Runtime/MRExtensionMethods_Vectors.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3eac0610ba2a24f40a0de36df2fe2323
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
